[Tarun] : Added cloudinary suppport - uploading files on cloudinary

+++ comment-service/src/main/java/com/flowboard/comment_service/repository/AttachmentRepository.java
@@ -7,10 +7,10 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AttachmentRepository extends JpaRepository<Attachment, Long> {
-    List<Attachment> findByCardId(Long cardId);
+public interface AttachmentRepository extends JpaRepository<Attachment, Integer> {
+    List<Attachment> findByCardId(Integer cardId);
 
-    Optional<Attachment> findByAttachmentId(Long attachmentId);
+    Optional<Attachment> findByAttachmentId(Integer attachmentId);
 
-    void deleteByAttachmentId(Long attachmentId);
+    void deleteByAttachmentId(Integer attachmentId);
 }
